Renamed framework & project to NodeMobile, make n-processors
