+++ tools/ios-framework/NodeMobile/Info.plist
