+++ tools/ios-framework/NodeMobile/NodeMobile.cpp
@@ -1,5 +1,5 @@
 
-#import "Node.h"
+#import "NodeMobile.h"
 
 namespace node {
     int Start(int argc, char *argv[]);
