+++ tools/ios_framework_prepare.sh
@@ -18,7 +18,7 @@ make clean
 TARGET_LIBRARY_PATH='tools/ios-framework/bin/arm64'
 
 ./configure --dest-os=ios --dest-cpu=arm64 --without-chakra-jit --enable-static --with-intl=none --openssl-no-asm
-make -j8
+make -j$(getconf _NPROCESSORS_ONLN)
 
 mkdir -p $TARGET_LIBRARY_PATH
 
@@ -37,7 +37,7 @@ make clean
 TARGET_LIBRARY_PATH='tools/ios-framework/bin/x64'
 
 ./configure --dest-os=ios --dest-cpu=x64 --without-chakra-jit --enable-static --with-intl=none --openssl-no-asm
-make -j8
+make -j$(getconf _NPROCESSORS_ONLN)
 
 mkdir -p $TARGET_LIBRARY_PATH
 
@@ -64,6 +64,7 @@ lipo -create "$TARGET_LIBRARY_PATH/arm64/libuv.a" "$TARGET_LIBRARY_PATH/x64/libu
 lipo -create "$TARGET_LIBRARY_PATH/arm64/libzlib.a" "$TARGET_LIBRARY_PATH/x64/libzlib.a" -output "$TARGET_LIBRARY_PATH/libzlib.a"
 
 #Create a path to build the frameworks into
+rm -rf out_ios
 mkdir -p out_ios
 cd out_ios
 FRAMEWORK_TARGET_DIR=${PWD}
@@ -71,10 +72,10 @@ cd ../
 
 NODELIB_PROJECT_PATH='tools/ios-framework'
 
-xcodebuild build -project $NODELIB_PROJECT_PATH/nodeLib.xcodeproj -target "Node" -configuration Release -arch arm64 -sdk "iphoneos" SYMROOT=$FRAMEWORK_TARGET_DIR
-xcodebuild build -project $NODELIB_PROJECT_PATH/nodeLib.xcodeproj -target "Node" -configuration Release -arch x86_64 -sdk "iphonesimulator" SYMROOT=$FRAMEWORK_TARGET_DIR
+xcodebuild build -project $NODELIB_PROJECT_PATH/NodeMobile.xcodeproj -target "NodeMobile" -configuration Release -arch arm64 -sdk "iphoneos" SYMROOT=$FRAMEWORK_TARGET_DIR
+xcodebuild build -project $NODELIB_PROJECT_PATH/NodeMobile.xcodeproj -target "NodeMobile" -configuration Release -arch x86_64 -sdk "iphonesimulator" SYMROOT=$FRAMEWORK_TARGET_DIR
 cp -RL $FRAMEWORK_TARGET_DIR/Release-iphoneos $FRAMEWORK_TARGET_DIR/Release-universal
-lipo -create $FRAMEWORK_TARGET_DIR/Release-iphoneos/Node.framework/Node $FRAMEWORK_TARGET_DIR/Release-iphonesimulator/Node.framework/Node -output $FRAMEWORK_TARGET_DIR/Release-universal/Node.framework/Node
+lipo -create $FRAMEWORK_TARGET_DIR/Release-iphoneos/NodeMobile.framework/NodeMobile $FRAMEWORK_TARGET_DIR/Release-iphonesimulator/NodeMobile.framework/NodeMobile -output $FRAMEWORK_TARGET_DIR/Release-universal/NodeMobile.framework/NodeMobile
 
 echo "Frameworks built to $FRAMEWORK_TARGET_DIR"
